Added dynamic game creation form
When a user wants to create a form there are different steps required
based on the game they are creating. This means that the HTML form must
be created dynamically based on the game. For this I have created a
javascript file which currently only supports Whist. Once the program
is fully functional for Whist, I will look to add support for other games.