Refactored new game menu
I have refactored the new game menu. It now has each of the three
required forms as part of the html file, with display set to none. When
the "game" select changes, the JavaScript automatically hides the
unneeded forms and displays the correct one. When the user presses
submit, the correct form is then submitted. This is much simpler and
clearer to read. Each select and input now also has a name so that it
can be accessed within the POST request. This is the problem that I ran
into and caused me to refactor the code, but it is probably for the best.

I have also added titles to each html page and will likely add more
metadata in the future as I get closer to the release date.