Added database
I have created an sqlite database with a table `games` that holds
data for each game. Currently it has a `gameID` and `config` as columns,
but I shall add more in the future. `gameID` acts as the primary key and
is a 6 digit integer. This would be shared between friends to join the
game. `config` is a text entry that holds JSON which will be used to
store the configuration options. This is because the data does not have
to parsed during SQL queries, only held and then passed to other parts
of the program (such as the backend or front end renderer).

Users can also add to the database by creating a new game and query it
by trying to join a game.