Added chatroom with websockets
I have added a chatroom option when creating a new game, which has no
configuration options. Once created, a player can join, send messages
and receive messages. This is achieved using websockets where the
message is sent to the server, which then sends it back to the user.
Currently this only works for a single user on local host and the next
step will be sending messages across the local network, and then between
any users on the internet.

Once the chatroom is finished, I will begin implementing the actual card
games. I will likely develop the back end, message format and front end
simultaneously, in small increments, so that each version is still a
functioning program.