Added CSS
I have added a CSS file, which adds some much needed styling to the
website. In general it has a dark brown background and white text, which
is occasionally inverted for emphasis (for example in the headers). Each
page has a header, which shows its title in the middle and a link to the
home page on the left.

Next I shall actually finish work on the websocket backend to allow for
users in the same local network to send messages to each other.