Messages can send across local network
Users can now create a chatroom and others on the local network can join
it. Once they have any messages sent will be broadcasted to all
connected clients. This is done by having a global variable linking a
gameID with all connected websockets. Creating a new game adds another
entry to the dictionary, which is updated when a join request is sent.
When a message is sent, it is then rebroadcasted to all connected
websockets.

I have also updated the UI so that a user needs to enter a username
before they join, which is sent alongside each message. The username and
a colon is then prepended to the message text before it is rendered.

Further improvements are still required for the UI and error handling for
KeyErrors when acessing the GAMES dict.