Fixed clipboard bug
I discovered a bug when attempting to host the website where
`navigator.clipboard` was undefined, which caused the webpage to break.
I discovered that this was because the browser had connected insecurely
over http and so blocked access to the clipboard. I have now added a
check for if the clipboard is undefined, which should resolve this issue.
This has reminded me that I need to add proper type hints and, more
importantly, error handling into my code.