Websocket server handles KeyErrors
Previously with restarting the websocket server in particular, the
database and `GAMES` global variable, which links a `gameID` to the set
of all websocket connections would become desynchronised. This would
cause clients to send requests with a `gameID` that is not in the
`GAMES` dictionary, causing a KeyError. The server side code now catches
this exception and logs an error. In the future I should also add some
feedback to the user that something has gone wrong.