Initial design for serialisation
I have made an initial design for the serialisation of messages, however
I have now realised that it will need to be iterated over and improved.
Therefore I am adding it to git so that previous versions and changes
can be tracked. I am also hoping that creating documentation can help me
to better design parts of the program and help others that wish to
develop or fork the program.