Reversed changes to functional state
I have reversed most of my changes to get back to a functional state. I
decided that removing that database was a poor decision and due to a
lack of development focus and good version history I have had to revert
a lot changes to get back to a functional state. I am hoping to use
some of them as a guideline for when I restart developement.