Server now fetches config from database
Previously the game config was sent alongside each message to the
websocket. However this was inefficient and prone to exploitation (it
would be easy for a user to change the config their browser sends, and
thus the behaviour of the server). Instead the config is now fetched
from the database when required, using Flask's wrappers to attempt to
reduce the number of concurrent connections.