Chatroom tracks sent messages
When a user creates a chatroom, it now tracks the messages they have
sent and when a new user joins, it sends all the previous messages to
them. This required some refactoring so that the handler functions are
called for all events including the create and join types. I also
created a global variable `CHATROOMS` that links a `gameID` to a list of
messages, which are then sent to the new user. I have also refactored
`chatroom.js` into a format I will likely copy for other game types.