Added userID cookie
I have added a `userID` cookie that can be used to identify users. I
plan for this can be used by the websocket server to identify a user,
rather than their websocket connection, allowing for websockets to be
replaced without breaking the game.

Currently the cookie has no expiration data so is a session cookie,
which means that it is destroyed when a user closes their browser. This
will likely be sufficient for a guest user, however I may have to change
things for users that have signed in. Rather than signing in every time
they close their browser, a user would likely prefer that they be
remembered. To do this I probably want a cookie that lasts for a longer
period of time (such as a month) and is replaced each time they visit
the website, though I should research this more.