userID now used to identify players
`ws_server.py` now uses `userID` to identify users rather than
websockets. `GAMES` now links from a `gameID` to a set of `userID`s and
`USERS` is a new variable that links a `userID` to a websocket. When a
new user joins, entries are added to both of these dictionaries. This is
because if the user reloads the page, a new websocket connection is
created which attempts to join the game. Previously this would cause 2
connections for the same player to be created, one of which no longer
worked, and this would cause errors. Now when the browser attempts to
rejoin, the server simply replaces the entry `USERS[userID]` with the
new websocket.