Username now linked to userID
A user's username is now linked to their `userID` rather than being sent
with each message. This means that reloading the page no longer requires
them to renter their username. I have also refactored large parts of
`server.py`, but I am not fully happy with it yet. I also need to add
security checks since currently a user can 'spoof' by entering another
user's `userID` and appear to send messages as them. This can be
prevented by checking that the message has been sent from the websocket
linked to the user and that user is part of the game they are sending a
message to.