Encapsulated GAMES global
I have tried to encapsulate the `GAMES` global variable, which links
from a `gameID` to a set of `userID`s, representing the users that have
connected to the game. Previously `ws_server.py` and other handlers
would directly access and change `GAMES` but I have now created getter
and setter functions for this. My argument for this is that it makes
creating a new handler easier and more stable. When writing a handler,
the programmer no longer has to know about internal implementation or
existence of `GAMES` making development easier. Their code is also
protected against changes in internal implementation, as long as the
interface does not change.