Whist UI mock-up
I have started to create the rendering logic for the front end, which
takes a JSON game state and generates HTML. In this initial mock-up I
have hard coded some HTML to see what the main view may look like. I am
planning to have a central table (ellipse) surrounded by players
(circles). Each object will have an accompanying text box that gives
information about it - either about the player (such as their bid) or
the game as a whole (such as the pot), but these have not been
implemented yet. I am hoping to have some of these elements in a
templates `div` with `display: none` that the javascript code can
reference. The renderer could then access each template by ID, copy it,
modify it and then add it to the main `<div id="content">`. This would
hopefully mitigate the amount of javascript that needs to be written as
the renderer would only have to modify the HTML rather than generate all
of it from scratch.