Added info boxes to Whist UI
I have added info boxes to the Whist UI, one for each player and a
general one for the table. Each player has an info box that shows
information specific to them (their username, bid, tricks won, etc.) and
the table info box shows information that is general to the game (the
trump suit, pot, etc.).

For each player their info box is a separate div that is positioned to
the right or below their seat. However the table does not have this and
paragraph tags are added directly to the table div. This is because I
want to have the information display on top of the table image (which is
now the `background-image` for the div) rather than beside it. The table
div also has some css set (`display: flex; align-items: center;
justify-content: center`) in the `styles.css` file that would otherwise
have to set using javascript. This is not difficult and benefit of
standardising info boxes between the players and table may be something
I revisit in the future, particularly since it could be quite strange /
confusing for another developer.

In general I will want to revisit all of the UI code so that I can
develop a generalised style (maybe even an internal library) that I can
copy for other games and reuse for different variations. However this
will require a good knowledge of the problem I am trying to solve and
the repeating patterns that occur - which I am unlikely to know until I
have created an initial solution.