Whist UI now draws each player's hand
I have updated the test game state in `whist.py` to send data about each
player's hand and `whist.js` to render the corresponding card. I have
added a `div class="hand"` to the player template that is populated with
the player's cards when a new game state is recieved. This required
changing some of the existing css and in its current state, the game
only renders correctly when in landscape on desktop. I should revisit
this when adding mobile support.

I have added 2 different types of card, one with the traditional images
(that is the default) and a simpler version (that may be better for
small images) and, in the future, I shall add an option to change
between them. The next steps are to allow for face-down cards, cards
drawn on the table and to render each user at the bottom of their
screen.