UI can render face down and community cards
In the JSON for a game state, a player's hand is a list of strings. For
example `[ "2H", "JD", "AS", "TC" ]` would be a hand containing the 2 of
Hearts, Jack of Diamonds, Ace of Spades and 10 of Clubs. However to
represent a hidden card an empty string is used; for example `[ "", "" ]`
represents a hand of 2 hidden cards. I have updated the UI to recognise
this and if it gets an empty string, to request the image for the card
back (`1B.svg`).

I have also added support for community cards that are placed on the
table. These are not used in whist but are common in poker and other
games and can be either face up or face down.