Different game state sent to each player
The server now generates a dummy game state and sends a copy to each player.
This is not an exact copy, instead a function takes the current game
state, a userID, and generates a partial copy of the game state. This
copy contains only the information that a player should have access to.
Currently this function is hard-coded but a key part of the modular
design of the program will be refactoring this so that it can be
generated based on game type and variations.