Current user shown at bottom
When playing a card game in real life, a user can see their own hand,
directly infront of them, and the rest of the table behind that. Most
digital versions of card games try to emulate this by placing the
current user at the bottom of the screen, with other players and the
table above them. I have decided to copy this format.

To achieve this, I have added a `current_user` attribute to the
`game_state` json sent to the user, which is an index into the existing
`players` list. The new function `rotatePlayers` can then use this index
to get what the current user is and rotate the list, in place, till that
user is at index 0. Other parts of the renderer can then use this fact
to display the players as they wish. For Whist, I have the current user
at the bottom and other players seated clockwise around the table.