Refactored handlers
I have refactored the handlers by moving the files to a separate
directory. I have also made substantial changes to `handlers/whist.py`
in how it 'censors' game states before sending them to users. Similar to
how I plan to design other parts of the program I have separated each
part of the process into separate procedures.

For example, the game state is mostly copied, but the `"players"`
attribute must be changed or 'censored'. To do this, a procedure is
called, passing the `game_state` and `censored_state`, that copies the
necessary data to the `censored_state`. Censoring each player is also a
series of smaller procedures for each key / value pair. This design is
supposed to be more flexibile as each function can be replaced without
affecting other parts of the program.