Waiting state now broadcast using game state code
When waiting for other users to join, connected users are sent a
`"waiting"` event. Previously this was hard coded and separate to how a
`"game_state"` is broadcast. However with the refactor I can now use the
`broadcast_game_state` method to broadcast both states. This gives me
more confidence that this is a good approach for this problem.