Fixed import error
I was previously experiencing an error where `server.py` would only run
if it imported the database with `from . import database` and
`ws_server.py` (which imports `server.py` itself) would only run if
`server.py` imported the database with `import database`. This issue was
resolved after I deleted `__pycache__/` and `__init__.py`.

Since python 3.3 `__init__.py` has not been necessary to mark a
directory as a module and in fact doing so has some quirks.
https://python-notes.curiousefficiency.org/en/latest/python_concepts/import_traps.html#the-init-py-trap
I am not running any initialisation code so for this reason I have
decided not to add the `__init__.py` file back.

I have also added a `pyproject.toml` file and installed the project in
an editable state with pip. This may also have fixed the issue, but did
not do so immediately. I will have to put more work into making the
project easy to ship and install (as well as writing a good README).