Renamed handlers/games to handlers/utils
I found the previous name `games` to be too vague and broad. I still do
not like the current name -- perhaps `ws_utils` would be more specific
-- so I may change it in the future. I also have to decide what to call
the code that manages the game state for each card game, which has
caused the naming collision and motivated the name change. This may also
be renamed in the future.