Initial game_state created and broadcast
When a new Whist game is created a corresponding `game_state` is
created, which is a dict that stores all data needed to run the game.
This includes a reference to the function that is called to run the
game. So far the default function can initialise a new game state and
broadcast this to the players. However this has revealed that the
current code for displaying a whist game state is not fit for purpose as
it cannot properly display each player with their starting hand of 13
cards. This means that the next step will have to be to fix this and
then I can continue to work on getting the game to run.