Cards now overlap
When displaying a player's hand their cards now overlap so that only the
indexes can be seen. This makes it much easier to display a large number
of cards, such as in Whist, where each player begins with 13 cards.

I achieved this using negative margins, which I calculate as a
percentage of the width that I set for the card. I choose this because I
wanted to use browser rendering as much as possible, rather than
something like canvas, or `position: absolute`.