Refactored `censor_game_state`
In Whist, the dealer deals the final card face up to set the trump suit.
Setting the trump suit based on this was easy to replicate but
displaying this card to all users was not possible with the design of
the current `censor_game_state` function. I have now refactored all code
so that it takes the entire `game_state`, even if it may not seem
necessary. This is so that if another variation requires access to
information that the default does not, it can read the entire
`game_state`. This allows every variation to share the same interface.

However, this required a refactor of how players are censored so each
attribute is censored for every player at one, rather each player being
censored in their entirety one at a time.