Moved `censor_game_state` to new file
I have moved `censor_game_state` to it's own file as it was getting too
large. I have also decided that it should not be called directly, but
should be generated based on the game's variations and the stage of the
game. I have not yet made this interface but that is the next step. I
will spend time refactoring it in a manner that I can hopefully copy for
the rest of the game code for when it comes to add variations to that.
For now it is still work in progress.