Fixed preserveAspectRatio
I have properly configured the preserveAspectRatio attribute of the
cards. Previously it was set to `"meet"` when it needed to have an
alignment set as well; now it is `"xMidYMid meet"` for all cards. I may
also want to look into using templating for reducing the repetition in
these files, but that is very low priority.