Users can play 1 trick
Users can now play a single trick with the UI showing the cards played
and incrementing the winner's 'Tricks Won' count. The back end now uses
and event-based architecture that waits for user input before
processing. This is not ideal or how I had designed the system, but it
seems to work. Next I will work on letting users play and entire hand of
Whist, after which I believe Whist will be close to completion.