Users can play hand of Whist
Users can now play an entire hand of Whist. I have also added a delay to
rendering a new `game_state` so that users can see the previous
`game_state` before it changes. When the last card is played in a trick,
the back end sends 2 `game_state`s back to back -- one for last card in
the trick and one for the beginning of the next trick. This delay
ensures that users can see the final card played before the next trick
begins.