Added Whist scoreboard
I have added a scoreboard that is shown to users after a hand of Whist.
Currently this ends the game, but next I am going to add proper win
conditions and a win screen. Afterwards I shall rewrite the rejoining
code so that reloading the page does not break the game. Once this is
complete, I shall add a couple of variations to Whist and the minimum
viable product will be complete.