Added first Whist variation
I have added the first variation to Whist, which is a change to the
scoring. In American Short Whist a game is played as the first
partnership to reach 7 odd tricks, rather than 5 as in British Short
Whist. I have created the necessary infrastructure to add any further
variations that I want and will now work on adding more to the game.